Increase argcnt size from uint8_t to uint64_t
While the VM is limited to 256 arguments, using APPLY+rest args are
unlimited, and this comes up in practice in fix-letrec when unioning
large sets.

+++ ir.h
@@ -22,7 +22,7 @@ typedef struct {
   bc *pc;
   uint16_t ir;
   uint16_t offset;
-  uint8_t argcnt;
+  uint32_t argcnt;
   snap_entry *slots;
 
   // Side trace info
